Corrected clear queues bug!! Happy 2015 :D
Solution was to update openvswitch package, now its with version 2.3.0 Dec 19 2014.
this version includes the "all" parameter in QoS management, which allows to clear all QoS records and all queues records at once
To update this, it is required to update python to 2.7 and also to perform a reload of the openvswitch kernel module, this can be done using:

depmod
modprobe openvswitch_mod
service openvswitch-switch start

The above commands should be used in case that there appears the bizarre error of not finding the interfaces associated with ov briges, for example:

in our configuration we have an entry in /etc/networ/interfaces called eth1br which is the "real" interface of one virtual machine,

also, this is an ov interface, this allows ov to directly controll the interfae functioning. For this to function, the ov kernel module should be running correctly.

Good luck!